fix: CI回帰2件 — 到達不能workflow_callをselector解決から除外しexecutionProfileをopencodeに限定する
