+++ src/infra/config/workflowCompanionResolution.ts
@@ -1,11 +1,10 @@
-import type {
-  ProviderRoutingEntry,
-  WorkflowConfig,
-  WorkflowStep,
-} from '../../core/models/index.js';
+import type { ProviderRoutingEntry, WorkflowConfig } from '../../core/models/index.js';
 import { isNormalAgentWorkflowStep } from '../../core/models/types.js';
 import type { CompiledProviderEnvironment } from './runtime-provider/environment.js';
-import { collectWorkflowCallSteps } from './loaders/workflowParallelTraversal.js';
+import {
+  collectReachableSteps,
+  collectReachableWorkflowCallSteps,
+} from './loaders/workflowParallelTraversal.js';
 import { resolveWorkflowCallTarget } from './loaders/workflowCallResolver.js';
 import { getWorkflowReference } from '../../core/workflow/workflow-reference.js';
 import { MAX_WORKFLOW_CALL_DEPTH } from '../../core/workflow/workflow-call-depth.js';
@@ -24,35 +23,6 @@ function collectLocalCompanionNames(workflow: WorkflowConfig): string[] {
     if (step.companion.moderator !== undefined) names.add(step.companion.moderator);
   }
   return [...names];
-}
-
-function collectReachableSteps(workflow: WorkflowConfig): WorkflowStep[] {
-  const stepsByName = new Map(workflow.steps.map((step) => [step.name, step]));
-  const visited = new Set<string>();
-  const pending = [workflow.initialStep];
-  const reachable: WorkflowStep[] = [];
-  while (pending.length > 0) {
-    const name = pending.shift();
-    if (name === undefined || visited.has(name)) continue;
-    visited.add(name);
-    const step = stepsByName.get(name);
-    if (step === undefined) continue;
-    reachable.push(step);
-    for (const next of step.rules?.map((rule) => rule.next) ?? []) {
-      if (next !== undefined && stepsByName.has(next) && !visited.has(next)) pending.push(next);
-    }
-    for (const monitor of workflow.loopMonitors ?? []) {
-      if (!monitor.cycle.includes(name)) continue;
-      for (const rule of monitor.judge.rules) {
-        if (stepsByName.has(rule.next) && !visited.has(rule.next)) pending.push(rule.next);
-      }
-    }
-  }
-  return reachable;
-}
-
-function collectReachableWorkflowCallSteps(workflow: WorkflowConfig) {
-  return collectWorkflowCallSteps(collectReachableSteps(workflow));
 }
 
 function collectCompanionNames(
